Use tokenState
