+++ frontend/src/State/states.ts
@@ -1,4 +1,4 @@
-import { atom, selector } from "recoil";
+import { atom } from "recoil";
 import {
   Division,
   Team,
@@ -8,9 +8,26 @@ import {
   Week,
 } from "./response-types";
 
+function getToken() {
+  try {
+    const item = window.localStorage.getItem("access_token");
+    if (item) {
+      const payload = JSON.parse(window.atob(item.split(".")[1]));
+      if (new Date(payload.exp * 1000) <= new Date()) {
+        window.localStorage.setItem("access_token", "");
+        return "";
+      }
+    }
+    return item ? item : "";
+  } catch (error) {
+    console.log(error);
+    return "";
+  }
+}
+
 export const tokenState = atom<string>({
   key: "accessToken",
-  default: "",
+  default: getToken(),
 });
 
 export const divisionsState = atom<Division[]>({
